Implement API for creating, listing, retrieving and patching simulations (#40)
The most important feature implemented in this PR is the ability to
create a simulation (`POST /api/v1/simulation/`). A simulation is
created based on the user settings. Once a simulation is created, it
cannot be deleted and only its status can be updated.

Additionally, authenticated API users can also:

- list simulations (`GET /api/v1/simulation/`)
- retrieve an individual simulation (`GET
/api/v1/simulation/{simulation_id}/`)
- update the status of an individual simulation (`PATCH
/api/v1/simulation/{simulation_id}/`)

**Note**. Currently security and data privacy is not enforced (users may
view and update each other's simulations).

Closes #30

---------

Signed-off-by: koosie0507 <3890788+koosie0507@users.noreply.github.com>