implement a database adapter for the executive sub-model (#45)
Use the settings stored in the database Instead of reading the settings
of the executive sub-model from a hardcoded config in the front-end.
Weights are generated randomly at every step. The aggrandisement
propensity of each minister is determined using the normal distribution
and an input probability which is parametrised at the level of the
government.